backend/foodgram/api/filters.py
from django_filters import rest_framework as filters

from app.models import Recipe


class RecipeFilter(filters.FilterSet):
    tags = filters.CharFilter(field_name='tags__slug')

    class Meta:
        model = Recipe
        fields = ['author', 'tags']
